:recycle: Split up TokenHistory tests
